[#660] mod: AppjamRankController로 컨벤션 맞춰서 책임 분리
