+++ src/main/java/org/sopt/app/presentation/appjamtamp/AppjamtampController.java
@@ -10,7 +10,6 @@ import java.util.Optional;
 import lombok.AllArgsConstructor;
 import lombok.val;
 
-import org.sopt.app.application.appjamrank.AppjamRankListInfo;
 import org.sopt.app.domain.enums.TeamNumber;
 import org.sopt.app.facade.AppjamtampFacade;
 import org.sopt.app.facade.MissionFacade;
@@ -61,19 +60,6 @@ public class AppjamtampController {
         val result = appjamtampFacade.getAppjamtamps(userId, findStampRequest.getMissionId(),
             findStampRequest.getNickname());
         val response = appjamtampResponseMapper.of(result);
-        return ResponseEntity.ok(response);
-    }
-
-    @Operation(summary = "앱잼 탬프 최근 인증한 팀 TOP3 조회하기")
-    @ApiResponses({
-        @ApiResponse(responseCode = "200", description = "success"),
-        @ApiResponse(responseCode = "500", description = "server error", content = @Content)
-    })
-    @GetMapping("/rank")
-    public ResponseEntity<AppjamtampResponse.AppjamtampRankListResponse> getRecentTeamRanks() {
-        AppjamRankListInfo appjamRankListInfo = appjamtampFacade.getRecentTeamRanks();
-        AppjamtampResponse.AppjamtampRankListResponse response = appjamtampResponseMapper.of(appjamRankListInfo);
-
         return ResponseEntity.ok(response);
     }
 }
